+++ wazo_auth/plugins/backends/ldap_user.py
@@ -14,18 +14,11 @@ logger = logging.getLogger(__name__)
 class LDAPUser(BaseAuthenticationBackend):
     def load(self, dependencies):
         super().load(dependencies)
-        config = dependencies['config']
         self._user_service = dependencies['user_service']
         self._group_service = dependencies['group_service']
         self._purposes = dependencies['purposes']
-        self.config = config['ldap']
-        self.uri = self.config['uri']
-        self.bind_dn = self.config.get('bind_dn', '')
-        self.bind_password = self.config.get('bind_password', '')
-        self.bind_anonymous = self.config.get('bind_anonymous', False)
-        self.user_base_dn = self.config['user_base_dn']
-        self.user_login_attribute = self.config['user_login_attribute']
-        self.user_email_attribute = self.config.get('user_email_attribute', 'mail')
+        self._ldap_service = dependencies['ldap_service']
+        self._tenant_service = dependencies['tenant_service']
 
     def get_acl(self, login, args):
         backend_acl = args.get('acl', [])
@@ -48,33 +41,76 @@ class LDAPUser(BaseAuthenticationBackend):
         return metadata
 
     def verify_password(self, username, password, args):
-        try:
-            wazo_ldap = _WazoLDAP(self.uri)
+        tenant_uuid = self._get_tenant(args['tenant_id'])['uuid']
 
-            if self.bind_anonymous or (self.bind_dn and self.bind_password):
-                if wazo_ldap.perform_bind(self.bind_dn, self.bind_password):
-                    user_dn = self._perform_search_dn(wazo_ldap, username)
+        config = self._get_ldap_config(tenant_uuid)
+        if not config:
+            logger.warning(
+                'Could not login: no LDAP config for tenant "%s"', tenant_uuid
+            )
+            return False
+
+        bind_dn = config.get('bind_dn')
+        bind_password = config.get('bind_password')
+        user_login_attribute = config.get('user_login_attribute')
+        user_email_attribute = config.get('user_email_attribute')
+        user_base_dn = config.get('user_base_dn')
+
+        wazo_ldap = _WazoLDAP(config)
+        try:
+            wazo_ldap.connect()
+            if bind_dn and bind_password:
+                if wazo_ldap.perform_bind(bind_dn, bind_password):
+                    user_dn = self._perform_search_dn(
+                        wazo_ldap,
+                        username,
+                        user_login_attribute,
+                        user_base_dn,
+                    )
                 else:
+                    logger.warning(
+                        'Could not login: service-level bind failed for "%s" on tenant "%s"',
+                        bind_dn,
+                        tenant_uuid,
+                    )
                     return False
             else:
-                user_dn = self._build_dn_with_config(username)
+                user_dn = self._build_dn_with_config(
+                    username, user_login_attribute, user_base_dn
+                )
 
             if not user_dn or not wazo_ldap.perform_bind(user_dn, password):
+                logger.debug(
+                    'Could not login: invalid credentials for user "%s" on tenant "%s"',
+                    user_dn,
+                    tenant_uuid,
+                )
                 return False
 
-            user_email = self._get_user_ldap_email(wazo_ldap, user_dn)
+            user_email = self._get_user_ldap_email(
+                wazo_ldap, user_dn, user_email_attribute
+            )
             if not user_email:
+                logger.debug(
+                    'Could not login: the LDAP user "%s" does not have an email address',
+                    user_dn,
+                )
                 return False
 
         except ldap.SERVER_DOWN:
-            logger.warning('LDAP : SERVER not responding on %s', self.uri)
+            logger.warning('LDAP : SERVER not responding on "%s"', wazo_ldap.uri)
             return False
         except ldap.LDAPError as exc:
-            logger.exception('ldap.LDAPError (%r, %r)', self.config, exc)
+            logger.exception('ldap.LDAPError (%r, %r)', config, exc)
             return False
 
-        pbx_user_uuid = self._get_user_uuid_by_ldap_attribute(user_email)
+        pbx_user_uuid = self._get_user_uuid_by_ldap_attribute(user_email, tenant_uuid)
         if not pbx_user_uuid:
+            logger.debug(
+                'Could not log in: user "%s" could not be found in tenant "%s"',
+                user_email,
+                tenant_uuid,
+            )
             return False
 
         args['pbx_user_uuid'] = pbx_user_uuid
@@ -82,58 +118,85 @@ class LDAPUser(BaseAuthenticationBackend):
 
         return True
 
-    def _get_user_uuid_by_ldap_attribute(self, user_email):
-        try:
-            user = next(iter(self._user_service.list_users(email_address=user_email)))
-        except StopIteration:
-            logger.warning(
-                '%s does not have an email associated with an auth user', user_email
-            )
-            return
-        return user['uuid']
+    def _get_ldap_config(self, tenant_uuid):
+        return self._ldap_service.get(tenant_uuid)
 
-    def _build_dn_with_config(self, login):
+    def _get_tenant(self, tenant_id):
+        return self._tenant_service.get_by_uuid_or_slug(None, tenant_id)
+
+    def _get_user_uuid_by_ldap_attribute(self, user_email, tenant_uuid):
+        for user in self._user_service.list_users(
+            email_address=user_email, tenant_uuid=tenant_uuid
+        ):
+            return user['uuid']
+
+        logger.warning(
+            '%s does not have an email associated with an auth user', user_email
+        )
+
+    def _build_dn_with_config(self, login, user_login_attribute, user_base_dn):
         login_esc = escape_dn_chars(login)
-        return '{}={},{}'.format(
-            self.user_login_attribute, login_esc, self.user_base_dn
-        )
+        return f'{user_login_attribute}={login_esc},{user_base_dn}'
 
-    def _get_user_ldap_email(self, wazo_ldap, user_dn):
+    def _get_user_ldap_email(self, wazo_ldap, user_dn, user_email_attribute):
         _, obj = wazo_ldap.perform_search(
-            user_dn, ldap.SCOPE_BASE, attrlist=[self.user_email_attribute]
+            user_dn, ldap.SCOPE_BASE, attrlist=[user_email_attribute]
         )
-        email = obj.get(self.user_email_attribute, None)
+        email = obj.get(user_email_attribute, None)
         email = email[0] if isinstance(email, list) else email
         if not email:
             logger.debug('LDAP : No email found for the user DN: %s', user_dn)
             return
         return email.decode('utf-8')
 
-    def _perform_search_dn(self, wazo_ldap, username):
+    def _perform_search_dn(
+        self, wazo_ldap, username, user_login_attribute, user_base_dn
+    ):
         username_esc = escape_filter_chars(username)
-        filterstr = '{}={}'.format(self.user_login_attribute, username_esc)
+        filterstr = f'{user_login_attribute}={username_esc}'
         dn, _ = wazo_ldap.perform_search(
-            self.user_base_dn, ldap.SCOPE_SUBTREE, filterstr=filterstr, attrlist=['']
+            user_base_dn, ldap.SCOPE_SUBTREE, filterstr=filterstr, attrlist=['']
         )
         if not dn:
             logger.debug(
                 'LDAP : No user DN for user_base dn: %s and filterstr: %s',
-                self.user_base_dn,
+                user_base_dn,
                 filterstr,
             )
         return dn
 
 
 class _WazoLDAP:
-    def __init__(self, uri):
-        self.uri = uri
-        self.ldapobj = self._create_ldap_obj(self.uri)
+    def __init__(self, config):
+        self.config = config
+        self.uri = self._build_uri(
+            config['protocol_security'], config['port'], config['host']
+        )
 
-    def _create_ldap_obj(self, uri):
-        ldapobj = ldap.initialize(uri)
+    def connect(self):
+        self.ldapobj = self._create_ldap_obj(self.config)
+
+    def _build_uri(self, protocol_security, port, host):
+        scheme = 'ldaps' if protocol_security == 'ldaps' else 'ldap'
+        return f'{scheme}://{host}:{port}'
+
+    def _create_ldap_obj(self, config):
+        ldapobj = ldap.initialize(self.uri)
         ldapobj.set_option(ldap.OPT_REFERRALS, 0)
         ldapobj.set_option(ldap.OPT_NETWORK_TIMEOUT, 2)
         ldapobj.set_option(ldap.OPT_TIMEOUT, 2)
+
+        version_map = {
+            2: ldap.VERSION2,
+            3: ldap.VERSION3,
+        }
+        ldapobj.set_option(
+            ldap.OPT_PROTOCOL_VERSION, version_map.get(config['protocol_version'], 3)
+        )
+
+        if config['protocol_security'] == 'tls':
+            ldapobj.set_option(ldap.OPT_X_TLS_NEWCTX, 0)
+            ldapobj.start_tls_s()
         return ldapobj
 
     def perform_bind(self, username, password):