+++ wazo_auth/plugins/backends/tests/test_ldap_user.py
@@ -27,6 +27,24 @@ class BaseTestCase(unittest.TestCase):
         self.group_service.get_acl = Mock()
         self.group_service.get_acl.return_value = ['acl2']
 
+        self.ldap_service = Mock()
+        self.ldap_service.get = Mock()
+        self.ldap_service.get.return_value = {
+            'protocol_security': '',
+            'protocol_version': 3,
+            'port': 389,
+            'host': 'host',
+            'user_base_dn': 'dc=example,dc=com',
+            'user_login_attribute': 'uid',
+            'user_email_attribute': 'mail',
+        }
+
+        self.tenant_service = Mock()
+        self.tenant_service.get_by_uuid_or_slug = Mock()
+        self.tenant_service.get_by_uuid_or_slug.return_value = {
+            'uuid': 'tenant-uuid-1234',
+        }
+
         user_metadata_plugin = Mock()
         user_metadata_plugin.get_token_metadata = Mock()
         user_metadata_plugin.get_token_metadata.return_value = {}
@@ -38,13 +56,6 @@ class BaseTestCase(unittest.TestCase):
 class TestGetACLS(BaseTestCase):
     def setUp(self):
         super().setUp()
-        config = {
-            'ldap': {
-                'uri': 'ldap://host:389',
-                'user_base_dn': 'dc=example,dc=com',
-                'user_login_attribute': 'uid',
-            },
-        }
         self.args = {
             'pbx_user_uuid': 'alice-uuid',
             'user_email': 'alice@wazo-auth.com',
@@ -53,9 +64,10 @@ class TestGetACLS(BaseTestCase):
         self.backend = LDAPUser()
         self.backend.load(
             {
-                'config': config,
                 'user_service': self.user_service,
                 'group_service': self.group_service,
+                'ldap_service': self.ldap_service,
+                'tenant_service': self.tenant_service,
                 'purposes': self.purposes,
             }
         )
@@ -73,20 +85,14 @@ class TestGetACLS(BaseTestCase):
 class TestGetMetadata(BaseTestCase):
     def setUp(self):
         super().setUp()
-        config = {
-            'ldap': {
-                'uri': 'ldap://host:389',
-                'user_base_dn': 'dc=example,dc=com',
-                'user_login_attribute': 'uid',
-            },
-        }
         self.args = {'pbx_user_uuid': 'alice-uuid', 'user_email': 'alice@wazo-auth.com'}
         self.backend = LDAPUser()
         self.backend.load(
             {
-                'config': config,
                 'user_service': self.user_service,
                 'group_service': self.group_service,
+                'ldap_service': self.ldap_service,
+                'tenant_service': self.tenant_service,
                 'purposes': self.purposes,
             },
         )
@@ -104,13 +110,6 @@ class TestGetMetadata(BaseTestCase):
 class TestVerifyPassword(BaseTestCase):
     def setUp(self):
         super().setUp()
-        self.config = {
-            'ldap': {
-                'uri': 'ldap://host:389',
-                'user_base_dn': 'dc=example,dc=com',
-                'user_login_attribute': 'uid',
-            },
-        }
         self.expected_user_dn = 'uid=foo,dc=example,dc=com'
         self.expected_user_email = 'foo@example.com'
         obj = {'mail': self.expected_user_email.encode('utf-8')}
@@ -122,14 +121,16 @@ class TestVerifyPassword(BaseTestCase):
         backend = LDAPUser()
         backend.load(
             {
-                'config': self.config,
                 'user_service': self.user_service,
                 'group_service': self.group_service,
+                'ldap_service': self.ldap_service,
+                'tenant_service': self.tenant_service,
                 'purposes': self.purposes,
             }
         )
-        wazo_ldap.side_effect = ldap.LDAPError
-        args = {}
+        wazo_ldap = wazo_ldap.return_value
+        wazo_ldap.connect.side_effect = ldap.LDAPError
+        args = {'tenant_id': 'test'}
 
         result = backend.verify_password('foo', 'bar', args)
         assert_that(result, equal_to(False))
@@ -138,14 +139,17 @@ class TestVerifyPassword(BaseTestCase):
         backend = LDAPUser()
         backend.load(
             {
-                'config': self.config,
                 'user_service': self.user_service,
                 'group_service': self.group_service,
+                'ldap_service': self.ldap_service,
+                'tenant_service': self.tenant_service,
                 'purposes': self.purposes,
             },
         )
-        wazo_ldap.side_effect = ldap.SERVER_DOWN
-        args = {}
+        wazo_ldap = wazo_ldap.return_value
+        wazo_ldap.uri = 'host'
+        wazo_ldap.connect.side_effect = ldap.SERVER_DOWN
+        args = {'tenant_id': 'test'}
 
         result = backend.verify_password('foo', 'bar', args)
         assert_that(result, equal_to(False))
@@ -154,9 +158,10 @@ class TestVerifyPassword(BaseTestCase):
         backend = LDAPUser()
         backend.load(
             {
-                'config': self.config,
                 'user_service': self.user_service,
                 'group_service': self.group_service,
+                'ldap_service': self.ldap_service,
+                'tenant_service': self.tenant_service,
                 'purposes': self.purposes,
             }
         )
@@ -165,7 +170,7 @@ class TestVerifyPassword(BaseTestCase):
         wazo_ldap.perform_bind.return_value = True
         wazo_ldap.perform_search.return_value = self.search_obj_result
         self.list_users.return_value = [{'uuid': 'alice-uuid'}]
-        args = {}
+        args = {'tenant_id': 'test'}
 
         result = backend.verify_password('foo', 'bar', args)
 
@@ -176,9 +181,10 @@ class TestVerifyPassword(BaseTestCase):
         backend = LDAPUser()
         backend.load(
             {
-                'config': self.config,
                 'user_service': self.user_service,
                 'group_service': self.group_service,
+                'ldap_service': self.ldap_service,
+                'tenant_service': self.tenant_service,
                 'purposes': self.purposes,
             }
         )
@@ -187,7 +193,7 @@ class TestVerifyPassword(BaseTestCase):
         wazo_ldap.perform_bind.return_value = True
         wazo_ldap.perform_search.return_value = ('uid=fo\\+o,dc=example,dc=com', Mock())
         self.list_users.return_value = [{'uuid': 'alice-uuid'}]
-        args = {}
+        args = {'tenant_id': 'test'}
 
         result = backend.verify_password('fo+o', 'bar', args)
 
@@ -197,14 +203,13 @@ class TestVerifyPassword(BaseTestCase):
         )
 
     def test_that_verify_password_escape_filter_chars(self, wazo_ldap):
-        extended_config = {'ldap': {'bind_anonymous': True}}
-        extended_config['ldap'].update(self.config['ldap'])
         backend = LDAPUser()
         backend.load(
             {
-                'config': self.config,
                 'user_service': self.user_service,
                 'group_service': self.group_service,
+                'ldap_service': self.ldap_service,
+                'tenant_service': self.tenant_service,
                 'purposes': self.purposes,
             }
         )
@@ -213,7 +218,7 @@ class TestVerifyPassword(BaseTestCase):
         wazo_ldap.perform_bind.return_value = True
         wazo_ldap.perform_search.return_value = ('uid=fo\\+o,dc=example,dc=com', Mock())
         self.list_users.return_value = [{'uuid': 'alice-uuid'}]
-        args = {}
+        args = {'tenant_id': 'test'}
 
         result = backend.verify_password('fo+o', 'bar', args)
 
@@ -226,15 +231,16 @@ class TestVerifyPassword(BaseTestCase):
         backend = LDAPUser()
         backend.load(
             {
-                'config': self.config,
                 'user_service': self.user_service,
                 'group_service': self.group_service,
+                'ldap_service': self.ldap_service,
+                'tenant_service': self.tenant_service,
                 'purposes': self.purposes,
             }
         )
         wazo_ldap = wazo_ldap.return_value
         wazo_ldap.perform_bind.return_value = False
-        args = {}
+        args = {'tenant_id': 'test'}
 
         result = backend.verify_password('foo', 'bar', args)
 
@@ -247,9 +253,10 @@ class TestVerifyPassword(BaseTestCase):
         backend = LDAPUser()
         backend.load(
             {
-                'config': self.config,
                 'user_service': self.user_service,
                 'group_service': self.group_service,
+                'ldap_service': self.ldap_service,
+                'tenant_service': self.tenant_service,
                 'purposes': self.purposes,
             }
         )
@@ -257,75 +264,24 @@ class TestVerifyPassword(BaseTestCase):
         wazo_ldap.perform_bind.return_value = True
         wazo_ldap.perform_search.return_value = self.search_obj_result
         self.list_users.return_value = []
-        args = {}
+        args = {'tenant_id': 'test'}
 
         result = backend.verify_password('foo', 'bar', args)
 
         assert_that(result, equal_to(False))
-        assert_that(args, equal_to({}))
-
-    def test_that_verify_password_calls_with_bind_anonymous(self, wazo_ldap):
-        extended_config = {'ldap': {'bind_anonymous': True}}
-        extended_config['ldap'].update(self.config['ldap'])
-        backend = LDAPUser()
-        backend.load(
-            {
-                'config': extended_config,
-                'user_service': self.user_service,
-                'group_service': self.group_service,
-                'purposes': self.purposes,
-            }
-        )
-        wazo_ldap = wazo_ldap.return_value
-        wazo_ldap.perform_bind.return_value = True
-        wazo_ldap.perform_search.return_value = self.search_obj_result
-        self.list_users.return_value = [{'uuid': 'alice-uuid'}]
-        args = {}
-
-        result = backend.verify_password('foo', 'bar', args)
-
-        assert_that(result, equal_to(True))
-        assert_that(
-            args,
-            equal_to({'pbx_user_uuid': 'alice-uuid', 'user_email': 'foo@example.com'}),
-        )
-        expected_call = [call('', ''), call(self.expected_user_dn, 'bar')]
-        wazo_ldap.perform_bind.assert_has_calls(expected_call)
-
-    def test_that_verify_password_calls_return_false_when_no_binding_with_anonymous(
-        self, wazo_ldap
-    ):
-        extended_config = {'ldap': {'bind_anonymous': True}}
-        extended_config['ldap'].update(self.config['ldap'])
-        wazo_ldap = wazo_ldap.return_value
-        backend = LDAPUser()
-        backend.load(
-            {
-                'config': extended_config,
-                'user_service': self.user_service,
-                'group_service': self.group_service,
-                'purposes': self.purposes,
-            }
-        )
-        wazo_ldap.perform_bind.return_value = False
-        args = {}
-
-        result = backend.verify_password('foo', 'bar', args)
-
-        assert_that(result, equal_to(False))
-        wazo_ldap.perform_bind.assert_called_once_with('', '')
+        assert_that(args, equal_to({'tenant_id': 'test'}))
 
     def test_that_verify_password_calls_with_bind_dn(self, wazo_ldap):
-        extended_config = {
-            'ldap': {'bind_dn': 'uid=foo,dc=example,dc=com', 'bind_password': 'S3cr$t'},
-        }
-        extended_config['ldap'].update(self.config['ldap'])
+        self.ldap_service.get.return_value.update(
+            {'bind_dn': 'uid=foo,dc=example,dc=com', 'bind_password': 'S3cr$t'}
+        )
         backend = LDAPUser()
         backend.load(
             {
-                'config': extended_config,
                 'user_service': self.user_service,
                 'group_service': self.group_service,
+                'ldap_service': self.ldap_service,
+                'tenant_service': self.tenant_service,
                 'purposes': self.purposes,
             }
         )
@@ -333,14 +289,20 @@ class TestVerifyPassword(BaseTestCase):
         wazo_ldap.perform_bind.return_value = True
         wazo_ldap.perform_search.return_value = self.search_obj_result
         self.list_users.return_value = [{'uuid': 'alice-uuid'}]
-        args = {}
+        args = {'tenant_id': 'test'}
 
         result = backend.verify_password('foo', 'bar', args)
 
         assert_that(result, equal_to(True))
         assert_that(
             args,
-            equal_to({'pbx_user_uuid': 'alice-uuid', 'user_email': 'foo@example.com'}),
+            equal_to(
+                {
+                    'tenant_id': 'test',
+                    'pbx_user_uuid': 'alice-uuid',
+                    'user_email': 'foo@example.com',
+                }
+            ),
         )
         expected_call = [
             call('uid=foo,dc=example,dc=com', 'S3cr$t'),
@@ -351,14 +313,16 @@ class TestVerifyPassword(BaseTestCase):
     def test_that_verify_password_calls_with_missing_bind_password_try_bind(
         self, wazo_ldap
     ):
-        extended_config = {'ldap': {'bind_dn': 'uid=foo,dc=example,dc=com'}}
-        extended_config['ldap'].update(self.config['ldap'])
+        self.ldap_service.get.return_value.update(
+            {'bind_dn': 'uid=foo,dc=example,dc=com'}
+        )
         backend = LDAPUser()
         backend.load(
             {
-                'config': extended_config,
                 'user_service': self.user_service,
                 'group_service': self.group_service,
+                'ldap_service': self.ldap_service,
+                'tenant_service': self.tenant_service,
                 'purposes': self.purposes,
             }
         )
@@ -366,14 +330,20 @@ class TestVerifyPassword(BaseTestCase):
         wazo_ldap.perform_bind.return_value = True
         wazo_ldap.perform_search.return_value = self.search_obj_result
         self.list_users.return_value = [{'uuid': 'alice-uuid'}]
-        args = {}
+        args = {'tenant_id': 'test'}
 
         result = backend.verify_password('foo', 'bar', args)
 
         assert_that(result, equal_to(True))
         assert_that(
             args,
-            equal_to({'pbx_user_uuid': 'alice-uuid', 'user_email': 'foo@example.com'}),
+            equal_to(
+                {
+                    'tenant_id': 'test',
+                    'pbx_user_uuid': 'alice-uuid',
+                    'user_email': 'foo@example.com',
+                }
+            ),
         )
         wazo_ldap.perform_bind.assert_called_once_with(self.expected_user_dn, 'bar')
 
@@ -381,7 +351,10 @@ class TestVerifyPassword(BaseTestCase):
 class TestWazoLDAP(unittest.TestCase):
     def setUp(self):
         self.config = {
-            'uri': 'ldap://host:389',
+            'protocol_security': '',
+            'protocol_version': 3,
+            'port': 389,
+            'host': 'host',
             'user_base_dn': 'dc=example,dc=com',
             'user_login_attribute': 'uid',
         }
@@ -390,16 +363,35 @@ class TestWazoLDAP(unittest.TestCase):
     def test_wazo_ldap_init(self, ldap_initialize):
         ldapobj = ldap_initialize.return_value = Mock()
 
-        _WazoLDAP(self.config['uri'])
+        wazo_ldap = _WazoLDAP(self.config)
+        wazo_ldap.connect()
 
-        ldap_initialize.assert_called_once_with(self.config['uri'])
+        ldap_initialize.assert_called_once_with(wazo_ldap.uri)
         ldapobj.set_option.assert_any_call(ldap.OPT_REFERRALS, 0)
         ldapobj.set_option.assert_any_call(ldap.OPT_NETWORK_TIMEOUT, 2)
         ldapobj.set_option.assert_any_call(ldap.OPT_TIMEOUT, 2)
+        ldapobj.set_option.assert_any_call(ldap.OPT_PROTOCOL_VERSION, ldap.VERSION3)
+
+    @patch('ldap.initialize')
+    def test_wazo_ldap_init_tls(self, ldap_initialize):
+        ldapobj = ldap_initialize.return_value = Mock()
+
+        with patch.dict(self.config, {'protocol_security': 'tls'}) as config:
+            wazo_ldap = _WazoLDAP(config)
+            wazo_ldap.connect()
+
+        ldap_initialize.assert_called_once_with(wazo_ldap.uri)
+        ldapobj.set_option.assert_any_call(ldap.OPT_REFERRALS, 0)
+        ldapobj.set_option.assert_any_call(ldap.OPT_NETWORK_TIMEOUT, 2)
+        ldapobj.set_option.assert_any_call(ldap.OPT_TIMEOUT, 2)
+        ldapobj.set_option.assert_any_call(ldap.OPT_PROTOCOL_VERSION, ldap.VERSION3)
+        ldapobj.set_option.assert_any_call(ldap.OPT_X_TLS_NEWCTX, 0)
+        ldapobj.start_tls_s.assert_called_once()
 
     @patch('ldap.initialize', Mock())
     def test_that_perform_bind(self):
         wazo_ldap = _WazoLDAP(self.config)
+        wazo_ldap.connect()
 
         result = wazo_ldap.perform_bind('username', 'password')
         self.assertEqual(result, True)
@@ -411,6 +403,7 @@ class TestWazoLDAP(unittest.TestCase):
         ldapobj = ldap_initialize.return_value = Mock()
 
         wazo_ldap = _WazoLDAP(self.config)
+        wazo_ldap.connect()
         ldapobj.simple_bind_s.side_effect = ldap.INVALID_CREDENTIALS()
         result = wazo_ldap.perform_bind('username', 'password')
         self.assertEqual(result, False)
@@ -419,6 +412,7 @@ class TestWazoLDAP(unittest.TestCase):
     def test_that_perform_search(self, ldap_initialize):
         ldapobj = ldap_initialize.return_value = Mock()
         wazo_ldap = _WazoLDAP(self.config)
+        wazo_ldap.connect()
         ldapobj.search_ext_s.return_value = ['result1']
 
         result = wazo_ldap.perform_search('base', 'scope')
@@ -430,6 +424,7 @@ class TestWazoLDAP(unittest.TestCase):
     ):
         ldapobj = ldap_initialize.return_value = Mock()
         wazo_ldap = _WazoLDAP(self.config)
+        wazo_ldap.connect()
         ldapobj.search_ext_s.side_effect = ldap.SIZELIMIT_EXCEEDED()
 
         result_dn, result_attr = wazo_ldap.perform_search('base', 'scope')
@@ -440,6 +435,7 @@ class TestWazoLDAP(unittest.TestCase):
     def test_that_perform_search_return_none_when_no_result(self, ldap_initialize):
         ldapobj = ldap_initialize.return_value = Mock()
         wazo_ldap = _WazoLDAP(self.config)
+        wazo_ldap.connect()
         ldapobj.search_ext_s.return_value = []
 
         result_dn, result_attr = wazo_ldap.perform_search('base', 'scope')